+++ CHANGES.md
@@ -1,6 +1,12 @@
 pkglogger changes
 =================
 
+4.1.0
+-----
+
+- Remove the TRACE log level as the same functionality can be achieved using the DEBUG level and the LOG_DEBUG (or DEBUG) envionment variable.
+- Added the latestLogFile getter.
+
 4.0.2
 -----